cli2/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "crowdpad-cli"
version = "0.1.0"
description = "Create a virtual joypad and listens for commands coming from the external websocket"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-uinput>=1.0.1",
]

[project.scripts]
start = "crowdpad_cli.main:main"

[tool.setuptools]
package-dir = {"" = "src"}